+++ CHANGES.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+Version 1.2.0
+-------------
+
+Unreleased
+
+- Cachelib is now used as backend. PR `#308 <https://github.com/pallets-eco/flask-caching/pull/308>`_.
+- Drop support for python 3.6
+
+
 Version 1.10.1
 --------------